# omics/omics_dashboard/static/swagger/swagger.yml
openapi: 3.0.0
info:
  description: This is used to control the data management aspect of Omics Dashboard
  version: '0.1'
  title: Omics Dashboard
  contact:
    email: foose.3@wright.edu
  license:
    name: GNU AGPL v3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
servers:
- url: https://birg.cs.wright.edu/omics/api
  description: Omics Dashboard @ BiRG
...
paths:
...
  /collections:
    get:
      tags:
      - Collections
      summary: Get a list of collections
      security:
      - bearerAuth: []
      responses:
        403:
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollectionOut'
  /collections/{id}:
    get:
      tags:
      - Collections
      summary: Get a specific collection
      parameters:
      - name: id
        in: path
        description: ID of collection
        required: true
        schema:
          type: integer
          format: int64
      security:
      - bearerAuth: []
      responses:
        403:
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionOut'
    post:
      tags:
      - Collections
      summary: Update a collection.
      parameters:
      - name: id
        in: path
        description: ID of collection
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: Data to update collection (all fields are optional and arbitrary fields can be added)
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordBase'
      security:
      - bearerAuth: []
      responses:
        404:
          description: Collection does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Not authorized to edit collection record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        200:
          description: User record updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
    delete:
      tags:
      - Collections
      summary: Delete a collection
      parameters:
      - name: id
        in: path
        description: ID of collection
        required: true
        schema:
          type: integer
          format: int64
      security:
      - bearerAuth: []
      responses:
        404:
          description: Collection does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Not authorized to edit collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        200:
          description: User record deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /collections/upload:
    post:
      tags:
      - Collections
      summary: Create a collection (from hdf5)
      requestBody:
        description: Collection metadata and file
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionSubmission'
      security:
      - bearerAuth: []
      responses:
        403:
          description: Not authorized to create collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionOut'
...
components:
  schemas:
...
    RecordBase:
      type: object
      properties:
        name:
          type: string
          description: The name of this record
        description:
          type: string
          description: The description of this record.
        groupPermissions:
          type: string
          description: The permissions allowed for members of the associated user group.
          enum: [full, readonly]
        allPermissions:
          type: string
          description: The permissions allowed by default for all users.
          enum: [full, readonly, none]
        userGroup:
          type: integer
          description: The user group id for the user group associated with this record
    Collection:
      allOf:
      - $ref: '#/components/schemas/RecordBase'
    CollectionOut:
      allOf:
      - $ref: '#/components/schemas/Collection'
      - properties:
          id:
            type: integer
            format: int64
            description: The ID of this record.
          datasets:
            type: array
            description: HDF5 Datasets (matrices) in collection.
            items:
              $ref: '#/components/schemas/Dataset'
          groups:
            type: array
            description: HDF5 Groups (like directories) in collection
            items:
              type: object
    CollectionSubmission:
      allOf:
      - $ref: '#/components/schemas/Collection'
      - type: object
        properties:
          file:
            type: string
            format: byte
            description: A base64-encoding of the HDF5 file
        required:
        - file
...
    Dataset:
      type: object
      description: An hdf5 dataset contained in a collection or sample.
      properties:
        nrows:
          description: The number of rows
          type: integer
        ncols:
          description: The number of columns
          type: integer
        path:
          description: The path to this dataset relative to the root group of the file
          type: string
        dtype:
          description: The Numpy datatype of the array.
          type: string
...
    InternalServerError:
      type: object
      description: A response object sent with status 500
      properties:
        message:
          type: string
          description: A basic description of the exception.
        traceback:
          type: string
          description: A python traceback of the exception.
    Error:
      type: object
      description: A response sent with an error status other than 500
      properties:
        message:
          type: string
          description: A basic description of the exception.
    Message:
      type: object
      description: A response sent when sending a data object does not make sense
      properties:
        message:
          type: string
          description: A basic description of the behavior.
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
- bearerAuth: []